added OpenAI preferences item

+++ OpenAI/Packages/com.openai.unity/Editor/OpenAIConfigurationSettingsInspector.cs
@@ -9,16 +9,49 @@ namespace OpenAI.Editor
     [CustomEditor(typeof(OpenAIConfigurationSettings))]
     internal class OpenAIConfigurationSettingsInspector : UnityEditor.Editor
     {
+        [SettingsProvider]
+        private static SettingsProvider Preferences()
+        {
+            return new SettingsProvider("Project/OpenAI", SettingsScope.Project, new[] { "OpenAI" })
+            {
+                label = "OpenAI",
+                guiHandler = OnPreferencesGui,
+                keywords = new[] { "OpenAI" }
+            };
+        }
+
+        private static void OnPreferencesGui(string searchContext)
+        {
+            var instance = GetOrCreateInstance();
+            var instanceEditor = CreateEditor(instance);
+            instanceEditor.OnInspectorGUI();
+        }
+
+        private static SerializedProperty apiKey;
+        private static SerializedProperty organization;
+
         private void OnEnable()
         {
-            var update = false;
-            var instance = target as OpenAIConfigurationSettings;
-            var currentPath = AssetDatabase.GetAssetPath(target);
+            GetOrCreateInstance(target);
+            apiKey = serializedObject.FindProperty(nameof(apiKey));
+            organization = serializedObject.FindProperty(nameof(organization));
+        }
 
-            if (string.IsNullOrWhiteSpace(currentPath))
-            {
-                return;
-            }
+        public override void OnInspectorGUI()
+        {
+            serializedObject.Update();
+            EditorGUILayout.Space();
+            EditorGUI.indentLevel++;
+            EditorGUILayout.PropertyField(apiKey);
+            EditorGUILayout.PropertyField(organization);
+            EditorGUI.indentLevel--;
+            serializedObject.ApplyModifiedProperties();
+        }
+
+        private static OpenAIConfigurationSettings GetOrCreateInstance(Object target = null)
+        {
+            var update = false;
+            OpenAIConfigurationSettings instance;
 
             if (!Directory.Exists("Assets/Resources"))
             {
@@ -26,24 +59,52 @@ namespace OpenAI.Editor
                 update = true;
             }
 
-            if (!currentPath.Contains("Resources"))
+            if (target != null)
             {
-                var newPath = $"Assets/Resources/{instance!.name}.asset";
+                instance = target as OpenAIConfigurationSettings;
 
-                if (!File.Exists(newPath))
+                var currentPath = AssetDatabase.GetAssetPath(instance);
+
+                if (string.IsNullOrWhiteSpace(currentPath))
                 {
-                    File.Move(Path.GetFullPath(currentPath), Path.GetFullPath(newPath));
-                    File.Move(Path.GetFullPath($"{currentPath}.meta"), Path.GetFullPath($"{newPath}.meta"));
+                    return instance;
                 }
-                else
+
+                if (!currentPath.Contains("Resources"))
                 {
-                    AssetDatabase.DeleteAsset(currentPath);
-                    var instances = AssetDatabase.FindAssets($"t:{nameof(OpenAIConfigurationSettings)}");
+                    var newPath = $"Assets/Resources/{instance!.name}.asset";
+
+                    if (!File.Exists(newPath))
+                    {
+                        File.Move(Path.GetFullPath(currentPath), Path.GetFullPath(newPath));
+                        File.Move(Path.GetFullPath($"{currentPath}.meta"), Path.GetFullPath($"{newPath}.meta"));
+                    }
+                    else
+                    {
+                        AssetDatabase.DeleteAsset(currentPath);
+                        var instances = AssetDatabase.FindAssets($"t:{nameof(OpenAIConfigurationSettings)}");
+                        var path = AssetDatabase.GUIDToAssetPath(instances[0]);
+                        instance = AssetDatabase.LoadAssetAtPath<OpenAIConfigurationSettings>(path);
+                    }
+
+                    update = true;
+                }
+            }
+            else
+            {
+                var instances = AssetDatabase.FindAssets($"t:{nameof(OpenAIConfigurationSettings)}");
+
+                if (instances.Length > 0)
+                {
                     var path = AssetDatabase.GUIDToAssetPath(instances[0]);
                     instance = AssetDatabase.LoadAssetAtPath<OpenAIConfigurationSettings>(path);
                 }
-
-                update = true;
+                else
+                {
+                    instance = CreateInstance<OpenAIConfigurationSettings>();
+                    AssetDatabase.CreateAsset(instance, $"Assets/Resources/{nameof(OpenAIConfigurationSettings)}.asset");
+                    update = true;
+                }
             }
 
             if (update)
@@ -54,6 +115,8 @@ namespace OpenAI.Editor
                     EditorGUIUtility.PingObject(instance);
                 };
             }
+
+            return instance;
         }
     }
 }
